Refactoring: move NDArrayWritable and transforms to datavec-api; remove datavec-nd4j-common entirely
