+++ datavec-api/src/main/java/org/datavec/api/transform/ndarray/NDArrayDistanceTransform.java
@@ -25,7 +25,7 @@ import org.datavec.api.transform.metadata.DoubleMetaData;
 import org.datavec.api.transform.schema.Schema;
 import org.datavec.api.writable.DoubleWritable;
 import org.datavec.api.writable.Writable;
-import org.datavec.common.data.NDArrayWritable;
+import org.datavec.api.writable.NDArrayWritable;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.ops.transforms.Transforms;
 import org.nd4j.shade.jackson.annotation.JsonProperty;
