+++ datavec-api/src/main/java/org/datavec/api/transform/ndarray/NDArrayMathFunctionTransform.java
@@ -17,13 +17,10 @@
 package org.datavec.api.transform.ndarray;
 
 import org.datavec.api.transform.MathFunction;
-import org.datavec.api.transform.Transform;
 import org.datavec.api.transform.metadata.ColumnMetaData;
 import org.datavec.api.transform.transform.BaseColumnTransform;
-import org.datavec.api.transform.transform.doubletransform.BaseDoubleTransform;
-import org.datavec.api.writable.DoubleWritable;
 import org.datavec.api.writable.Writable;
-import org.datavec.common.data.NDArrayWritable;
+import org.datavec.api.writable.NDArrayWritable;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.factory.Nd4j;
 import org.nd4j.linalg.ops.transforms.Transforms;
