+++ datavec-api/src/main/java/org/datavec/api/util/ndarray/RecordConverter.java
@@ -1,5 +1,5 @@
-/*-
- *  * Copyright 2016 Skymind, Inc.
+/*
+ *  * Copyright 2017 Skymind, Inc.
  *  *
  *  *    Licensed under the Apache License, Version 2.0 (the "License");
  *  *    you may not use this file except in compliance with the License.
@@ -14,12 +14,12 @@
  *  *    limitations under the License.
  */
 
-package org.datavec.common;
+package org.datavec.api.util.ndarray;
 
 import org.datavec.api.writable.DoubleWritable;
 import org.datavec.api.writable.IntWritable;
 import org.datavec.api.writable.Writable;
-import org.datavec.common.data.NDArrayWritable;
+import org.datavec.api.writable.NDArrayWritable;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.dataset.api.DataSet;
 import org.nd4j.linalg.factory.Nd4j;
