+++ datavec-data/datavec-data-image/src/main/java/org/datavec/image/recordreader/BaseImageRecordReader.java
@@ -28,8 +28,8 @@ import org.datavec.api.util.files.FileFromPathIterator;
 import org.datavec.api.util.files.URIUtil;
 import org.datavec.api.writable.IntWritable;
 import org.datavec.api.writable.Writable;
-import org.datavec.common.RecordConverter;
-import org.datavec.common.data.NDArrayWritable;
+import org.datavec.api.util.ndarray.RecordConverter;
+import org.datavec.api.writable.NDArrayWritable;
 import org.datavec.image.loader.BaseImageLoader;
 import org.datavec.image.loader.ImageLoader;
 import org.datavec.image.loader.NativeImageLoader;
