+++ datavec-local/src/test/java/org/datavec/local/transforms/TableRecordsTest.java
@@ -3,7 +3,7 @@ package org.datavec.local.transforms;
 import org.datavec.api.transform.schema.Schema;
 import org.datavec.api.writable.DoubleWritable;
 import org.datavec.api.writable.Writable;
-import org.datavec.common.RecordConverter;
+import org.datavec.api.util.ndarray.RecordConverter;
 import org.datavec.dataframe.api.Table;
 import org.junit.Test;
 import org.nd4j.linalg.api.ndarray.INDArray;
