+++ datavec-spark/src/main/java/org/datavec/spark/transform/analysis/columns/NDArrayAnalysisCounter.java
@@ -18,7 +18,7 @@ package org.datavec.spark.transform.analysis.columns;
 
 import org.datavec.api.transform.analysis.columns.NDArrayAnalysis;
 import org.datavec.api.writable.Writable;
-import org.datavec.common.data.NDArrayWritable;
+import org.datavec.api.writable.NDArrayWritable;
 import org.datavec.spark.transform.analysis.AnalysisCounter;
 import org.nd4j.linalg.api.ndarray.INDArray;
