+++ datavec-spark/src/main/java/org/datavec/spark/transform/analysis/histogram/NDArrayHistogramCounter.java
@@ -18,7 +18,7 @@ package org.datavec.spark.transform.analysis.histogram;
 
 import org.datavec.api.writable.DoubleWritable;
 import org.datavec.api.writable.Writable;
-import org.datavec.common.data.NDArrayWritable;
+import org.datavec.api.writable.NDArrayWritable;
 import org.nd4j.linalg.api.ndarray.INDArray;
 
 /**
