+++ datavec-spark/src/test/java/org/datavec/spark/transform/NormalizationTests.java
@@ -5,7 +5,7 @@ import org.apache.spark.sql.Row;
 import org.datavec.api.transform.schema.Schema;
 import org.datavec.api.writable.DoubleWritable;
 import org.datavec.api.writable.Writable;
-import org.datavec.common.RecordConverter;
+import org.datavec.api.util.ndarray.RecordConverter;
 import org.datavec.spark.BaseSparkTest;
 import org.junit.Test;
 import org.nd4j.linalg.api.ndarray.INDArray;
